Adaptación de script PostgreSQL

+++ scripts/bbdd/3.5/postgresql/sistra_bantel_update_from_3.5.14_to_3.5.15.sql
@@ -1,4 +1,4 @@
-alter table BTE_PROAPL  add   TAP_IDESIA           NUMBER(20);
+alter table BTE_PROAPL  add   TAP_IDESIA           bigint;
 
 comment on column BTE_PROAPL.TAP_IDESIA is
 'IDENTIFICADOR SIA';
